+++ ViewModels/OverrideLogs/ORLog/ORLogCostViewModel.cs
@@ -7,7 +7,7 @@ namespace ViewModels.OverrideLogs.ORLog
     {
         public long Id { get; set; }
         [Display(Name = "Hours")]
-        public int OverrideHours { get; set; }
+        public double OverrideHours { get; set; }
 
         public int HeadCount { get; set; }
         public CraftSkillBriefViewModel CraftSkill { get; set; } = new CraftSkillBriefViewModel();
